@@ -12,15 +12,21 @@
   "license": "MIT",
   "dependencies": {
     "@prisma/client": "^5.6.0",
-    "next": "12.0.10",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-markdown": "8.0.0"
+    "next": "14.0.3",
+    "react": "^18",
+    "react-dom": "^18",
+    "react-markdown": "^9.0.1"
   },
   "devDependencies": {
-    "@types/node": "17.0.14",
-    "@types/react": "17.0.38",
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "autoprefixer": "^10.0.1",
+    "eslint": "^8",
+    "eslint-config-next": "14.0.3",
+    "postcss": "^8",
     "prisma": "^5.6.0",
-    "typescript": "4.5.5"
+    "tailwindcss": "^3.3.0",
+    "typescript": "^5"
   }
 }
