@@ -1,54 +1,33 @@
-import React from "react";
-import Link from "next/link";
-import { useRouter } from "next/router";
+'use client';
+import React from 'react';
+import Link from 'next/link';
+// import { useRouter } from 'next/router';
+import { usePathname } from 'next/navigation';
+import styles from './header.module.css';
 
 const Header: React.FC = () => {
-  const router = useRouter();
-  const isActive: (pathname: string) => boolean = (pathname) =>
-    router.pathname === pathname;
+  // const router = useRouter();
+  const pathname = usePathname();
+  const isActive: (pathname: string) => boolean = (currentPathname) =>
+    // router.pathname === pathname;
+    pathname === currentPathname;
 
   let left = (
-    <div className="left">
-      <Link href="/">
-        <a className="bold" data-active={isActive("/")}>
+    <div className={styles.left}>
+      <Link href='/' legacyBehavior>
+        <a className={styles.bold} data-active={isActive('/')}>
           Feed
         </a>
       </Link>
-      <style jsx>{`
-        .bold {
-          font-weight: bold;
-        }
-
-        a {
-          text-decoration: none;
-          color: #000;
-          display: inline-block;
-        }
-
-        .left a[data-active="true"] {
-          color: gray;
-        }
-
-        a + a {
-          margin-left: 1rem;
-        }
-      `}</style>
     </div>
   );
 
   let right = null;
 
   return (
-    <nav>
+    <nav className={styles.nav}>
       {left}
       {right}
-      <style jsx>{`
-        nav {
-          display: flex;
-          padding: 2rem;
-          align-items: center;
-        }
-      `}</style>
     </nav>
   );
 };
