@@ -1,6 +1,7 @@
-import React from "react";
-import Router from "next/router";
-import ReactMarkdown from "react-markdown";
+import React from 'react';
+import ReactMarkdown from 'react-markdown';
+import Link from 'next/link';
+import styles from './post.module.css';
 
 export type PostProps = {
   id: string;
@@ -14,18 +15,14 @@ export type PostProps = {
 };
 
 const Post: React.FC<{ post: PostProps }> = ({ post }) => {
-  const authorName = post.author ? post.author.name : "Unknown author";
+  const authorName = post.author ? post.author.name : 'Unknown author';
   return (
-    <div onClick={() => Router.push("/p/[id]", `/p/${post.id}`)}>
-      <h2>{post.title}</h2>
-      <small>By {authorName}</small>
-      <ReactMarkdown children={post.content} />
-      <style jsx>{`
-        div {
-          color: inherit;
-          padding: 2rem;
-        }
-      `}</style>
+    <div className={styles.post}>
+      <Link href={post.id}>
+        <h2>{post.title}</h2>
+        <small>By {authorName}</small>
+        <ReactMarkdown children={post.content} />
+      </Link>
     </div>
   );
 };
